backend/requirements/base.txt
# Core Django Framework
Django>=4.2.7,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1

# Database
psycopg2-binary>=2.9.7

# Environment Management
python-decouple>=3.8

# Authentication & Authorization
djangorestframework-simplejwt>=5.3.0
django-oauth-toolkit>=1.7.1
social-auth-app-django>=5.4.0

# API Documentation
drf-spectacular>=0.26.5

# Caching & Performance
django-redis>=5.4.0
redis>=5.0.1

# Background Tasks
celery>=5.3.4
django-celery-beat>=2.5.0

# Image Processing (for avatars)
Pillow>=10.0.1

# HTTP Requests (for TMDb API)
requests>=2.31.0

# Timezone Support
pytz>=2023.3

# JSON Web Tokens
PyJWT>=2.8.0

# Date/Time utilities
python-dateutil>=2.8.2
